增加视频标签列表查询功能

+++ bliliblili-server/src/main/java/com/bliliblili/service/impl/VideoTagServiceImpl.java
@@ -1,10 +1,13 @@
 package com.bliliblili.service.impl;
 
+import com.bliliblili.dao.VideoDao;
 import com.bliliblili.dao.VideoTagDao;
 import com.bliliblili.domain.entity.Tag;
 import com.bliliblili.service.VideoTagService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.Date;
 
 /**
  * @ author ljw
@@ -16,18 +19,21 @@ public class VideoTagServiceImpl implements VideoTagService {
     @Autowired
     private VideoTagDao videoTagDao;
 
-
     public Long addVideoTag(Tag tag) {
         Tag dbTag = videoTagDao.getTagByName(tag.getName());
-        if(dbTag!= null){
+        if (dbTag != null) {
             return dbTag.getId();
         }
+        tag.setCreateTime(new Date());
         videoTagDao.addVideoTag(tag);
         return tag.getId();
     }
 
 
     public void deleteVideoTag(Long tagId) {
-        videoTagDao.deleteVideoTag(tagId);
+        Integer count = videoTagDao.getVideoByTagId(tagId);
+        if(count != null){
+            videoTagDao.deleteVideoTag(tagId);
+        }
     }
 }
